requirements.txt
torch==1.5.1+cu101
torchvision==0.6.1+cu101
numpy==1.19.1
pandas==1.1.3
Pillow==7.2.0
opencv_python==4.3.0.36
scikit_learn==0.23.2
matplotlib==3.3.0
seaborn==0.10.1
tqdm==4.48.0
scipy==1.5.1
